fix: Commented out getAccessControlResult and related test
This had previously been commented out and there is
a TODO message to evaluate whether it's needed.
As there was still a test for this, I has uncommented the code so
the test would pass, however, this actually breaks the PUT form
submissions, so I have now recommened the code and also commented
out the test.