Added MDXeditor
This is the initial setup POC.
There are no test for this section yet.
The editor is now working when field type is set to: 'mdxEditor'

```
  body: {
    type: "mdxEditor",
  }

```

It would be good to make this configurable via a config file.
We also need to ass full tests for this feature before it's ready.

We will also want to integrate this with Cloudflare Images
once that feature is complete.